+++ src/core/datatypes/complex/container/Container.h
@@ -71,7 +71,7 @@ class Container : public RbComplex {
         // Container functions and basic utility functions you should not have to override
         ContainerIterator               begin(void) const;                                                  //!< Begin iterator
         ContainerIterator               end(void) const;                                                    //!< End iterator
-        int                             getDim(void) const { return int(length.size()); }                   //!< Get number of dimensions (1 for vector, 2 for matrix, etc)
+        size_t                          getDim(void) const { return length.size(); }                        //!< Get number of dimensions (1 for vector, 2 for matrix, etc)
         const std::string&              getElementType(void) const { return elementType; }                  //!< Get element type (this method returns a string because containers of containers are not allowed so dim of elements is always 0)
         size_t                          getIndexOfName(size_t k, const std::string& s) const;               //!< Get index in dimension k corresponding to name s
         VectorInteger                   getIntegerIndex(const VectorIndex& index) const;                    //!< Get integer index to a single container element or to a subcontainer from generic index
